+++ apps/openmw-mp/Script/Functions/World.cpp
@@ -69,7 +69,7 @@ int WorldFunctions::GetObjectCharge(unsigned int i) noexcept
     return readEvent->worldObjects.at(i).charge;
 }
 
-int WorldFunctions::GetObjectEnchantmentCharge(unsigned int i) noexcept
+double WorldFunctions::GetObjectEnchantmentCharge(unsigned int i) noexcept
 {
     return readEvent->worldObjects.at(i).enchantmentCharge;
 }
@@ -152,7 +152,7 @@ int WorldFunctions::GetContainerItemCharge(unsigned int objectIndex, unsigned in
         .containerItems.at(itemIndex).charge;
 }
 
-int WorldFunctions::GetContainerItemEnchantmentCharge(unsigned int objectIndex, unsigned int itemIndex) noexcept
+double WorldFunctions::GetContainerItemEnchantmentCharge(unsigned int objectIndex, unsigned int itemIndex) noexcept
 {
     return readEvent->worldObjects.at(objectIndex)
         .containerItems.at(itemIndex).enchantmentCharge;
@@ -204,7 +204,7 @@ void WorldFunctions::SetObjectCharge(int charge) noexcept
     tempWorldObject.charge = charge;
 }
 
-void WorldFunctions::SetObjectEnchantmentCharge(int enchantmentCharge) noexcept
+void WorldFunctions::SetObjectEnchantmentCharge(double enchantmentCharge) noexcept
 {
     tempWorldObject.enchantmentCharge = enchantmentCharge;
 }
@@ -282,7 +282,7 @@ void WorldFunctions::SetContainerItemCharge(int charge) noexcept
     tempContainerItem.charge = charge;
 }
 
-void WorldFunctions::SetContainerItemEnchantmentCharge(int enchantmentCharge) noexcept
+void WorldFunctions::SetContainerItemEnchantmentCharge(double enchantmentCharge) noexcept
 {
     tempContainerItem.enchantmentCharge = enchantmentCharge;
 }
