+++ components/openmw-mp/Base/BaseStructs.hpp
@@ -12,7 +12,7 @@ namespace mwmp
         std::string refId;
         int count;
         int charge;
-        int enchantmentCharge;
+        double enchantmentCharge;
 
         inline bool operator==(const Item& rhs)
         {
